+++ test/Service/StorageCacheAbstractServiceFactoryTest.php
@@ -10,10 +10,12 @@
 namespace ZendTest\Cache\Service;
 
 use Zend\Cache;
+use Zend\ServiceManager\Config;
 use Zend\ServiceManager\ServiceManager;
 
 /**
  * @group      Zend_Cache
+ * @covers Zend\Cache\StorageFactory<extended>
  */
 class StorageCacheAbstractServiceFactoryTest extends \PHPUnit_Framework_TestCase
 {
@@ -23,18 +25,34 @@ class StorageCacheAbstractServiceFactoryTest extends \PHPUnit_Framework_TestCase
     {
         Cache\StorageFactory::resetAdapterPluginManager();
         Cache\StorageFactory::resetPluginManager();
+        $config = [
+            'services' => [
+                'config' => [
+                    'caches' => [
+                        'Memory' => [
+                            'adapter' => 'Memory',
+                            'plugins' => ['Serializer', 'ClearExpiredByFactor'],
+                        ],
+                        'Foo' => [
+                            'adapter' => 'Memory',
+                            'plugins' => ['Serializer', 'ClearExpiredByFactor'],
+                        ],
+                    ]
+                ],
+            ],
+            'abstract_factories' => [
+                'Zend\Cache\Service\StorageCacheAbstractServiceFactory'
+            ]
+        ];
         $this->sm = new ServiceManager();
-        $this->sm->setService('Config', ['caches' => [
-            'Memory' => [
-                'adapter' => 'Memory',
-                'plugins' => ['Serializer', 'ClearExpiredByFactor'],
-            ],
-            'Foo' => [
-                'adapter' => 'Memory',
-                'plugins' => ['Serializer', 'ClearExpiredByFactor'],
-            ],
-        ]]);
-        $this->sm->addAbstractFactory('Zend\Cache\Service\StorageCacheAbstractServiceFactory');
+        if (method_exists($this->sm, 'configure')) {
+            // v3
+            $this->sm->configure($config);
+        } else {
+            // v2
+            $config = new Config($config);
+            $config->configureServiceManager($this->sm);
+        }
     }
 
     public function tearDown()