+++ test/Service/StorageCacheFactoryTest.php
@@ -10,10 +10,12 @@
 namespace ZendTest\Cache\Service;
 
 use Zend\Cache;
+use Zend\ServiceManager\Config;
 use Zend\ServiceManager\ServiceManager;
 
 /**
  * @group      Zend_Cache
+ * @covers Zend\Cache\Service\StorageCacheFactory
  */
 class StorageCacheFactoryTest extends \PHPUnit_Framework_TestCase
 {
@@ -23,12 +25,28 @@ class StorageCacheFactoryTest extends \PHPUnit_Framework_TestCase
     {
         Cache\StorageFactory::resetAdapterPluginManager();
         Cache\StorageFactory::resetPluginManager();
+        $config = [
+            'services' => [
+                'config' => [
+                    'cache' => [
+                        'adapter' => 'Memory',
+                        'plugins' => ['Serializer', 'ClearExpiredByFactor'],
+                    ]
+                ]
+            ],
+            'factories' => [
+                'CacheFactory' => \Zend\Cache\Service\StorageCacheFactory::class
+            ]
+        ];
         $this->sm = new ServiceManager();
-        $this->sm->setService('Config', ['cache' => [
-            'adapter' => 'Memory',
-            'plugins' => ['Serializer', 'ClearExpiredByFactor'],
-        ]]);
-        $this->sm->setFactory('CacheFactory', 'Zend\Cache\Service\StorageCacheFactory');
+        if (method_exists($this->sm, 'configure')) {
+            // v3
+            $this->sm->configure($config);
+        } else {
+            // v2
+            $config = new Config($config);
+            $config->configureServiceManager($this->sm);
+        }
     }
 
     public function tearDown()