+++ test/Storage/Adapter/MemoryTest.php
@@ -13,6 +13,7 @@ use Zend\Cache;
 
 /**
  * @group      Zend_Cache
+ * @covers Zend\Cache\Storage\Adapter\Memory<extended>
  */
 class MemoryTest extends CommonAdapterTest
 {