+++ test/Storage/Adapter/RedisResourceManagerTest.php
@@ -17,6 +17,7 @@ use Zend\Cache\Storage\Adapter\RedisResourceManager;
 
 /**
  * @group      Zend_Cache
+ * @covers Zend\Cache\Storage\Adapter\RedisResourceManager
  */
 class RedisResourceManagerTest extends \PHPUnit_Framework_TestCase
 {