+++ test/Storage/Adapter/SessionTest.php
@@ -14,11 +14,20 @@ use Zend\Session\Container as SessionContainer;
 
 /**
  * @group      Zend_Cache
+ * @covers Zend\Cache\Storage\Adapter\Session<extended>
  */
 class SessionTest extends CommonAdapterTest
 {
     public function setUp()
     {
+        if (! class_exists(SessionContainer::class)) {
+            $this->markTestSkipped(
+                'Skipping zend-session-related tests until that component is '
+                . 'forwards-compatible with zend-stdlib, zend-servicemanager, '
+                . 'and zend-eventmanager v3'
+            );
+        }
+
         $_SESSION = [];
         SessionContainer::setDefaultManager(null);
         $sessionContainer = new SessionContainer('Default');
@@ -34,6 +43,10 @@ class SessionTest extends CommonAdapterTest
 
     public function tearDown()
     {
+        if (! class_exists(SessionContainer::class)) {
+            return;
+        }
+
         $_SESSION = [];
         SessionContainer::setDefaultManager(null);
     }